#38
- http-server/middleware/log: модуль удален (функционал перенесен в preset/node) (major)
